+++ packages/typespec-rust/src/shared/shared.ts
@@ -89,7 +89,6 @@ function unwrap(type: rust.Type): rust.Type | undefined {
     case 'literal':
       return type.valueKind;
     case 'option':
-    case 'pageIterator':
     case 'pager':
     case 'ref':
       return type.type;