+++ packages/typespec-rust/CHANGELOG.md
@@ -11,6 +11,7 @@
 * The algorithm for renaming of pageable methods has changed which can cause some method names to change.
 * Some method parameters are now borrowed instead of owned.
 * Use helper `UrlExt::append_path()` from `azure_core` when constructing the request's `Url`.
+* Pageable methods always return a `Pager<T>` instead of sometimes returning a `PageIterator<T>`.
 
 ### Bugs Fixed